requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy<2.0
setuptools
git+https://github.com/ageitgey/face_recognition_models

# Configuration
PyYAML>=6.0

# Web dashboard
Flask>=3.0

# Testing
pytest>=7.0

# Optional: Excel export (install if needed)
# openpyxl>=3.1

# Optional: FAISS for fast nearest-neighbor (large face databases)
# faiss-cpu>=1.7
